app/core/encryption.py
"""
Encryption Service for Vaulto Note

This module will handle encryption/decryption of note content.
Currently contains placeholder functions that pass through data unchanged.

ENCRYPTION CONTRACT:
-------------------
1. All note content MUST be encrypted before storage in DB
2. Encryption happens in this service layer, NOT in routes
3. API accepts/returns plaintext, DB stores encrypted data

FIELDS:
-------
- title: Plaintext (for search/indexing)
- encrypted_title: Encrypted title (optional, for future full encryption)
- content: LEGACY plaintext field (will be removed)
- encrypted_content: Encrypted content (current standard)

FUTURE IMPLEMENTATION:
---------------------
Replace placeholder functions with actual encryption using:
- Fernet (symmetric encryption)
- User-specific keys (derived from password or separate encryption key)
- Key derivation: PBKDF2 or Argon2

USAGE:
------
from app.core.encryption import encrypt_content, decrypt_content

# In route handler:
encrypted = encrypt_content(plaintext_content, user_key)
# Store encrypted in DB

# When retrieving:
plaintext = decrypt_content(encrypted_content, user_key)
# Return plaintext to API
"""

from typing import Optional


def encrypt_content(plaintext: str, encryption_key: Optional[str] = None) -> str:
    """
    Encrypt note content.
    
    Args:
        plaintext: The plaintext content to encrypt
        encryption_key: User-specific encryption key (optional for now)
    
    Returns:
        Encrypted content as base64 string
    
    TODO: Implement actual encryption using Fernet or similar
    For now, this is a passthrough function.
    """
    # PLACEHOLDER: Return plaintext unchanged
    # Future: Use Fernet(key).encrypt(plaintext.encode()).decode()
    return plaintext


def decrypt_content(encrypted: str, encryption_key: Optional[str] = None) -> str:
    """
    Decrypt note content.
    
    Args:
        encrypted: The encrypted content (base64 string)
        encryption_key: User-specific encryption key (optional for now)
    
    Returns:
        Decrypted plaintext content
    
    TODO: Implement actual decryption using Fernet or similar
    For now, this is a passthrough function.
    """
    # PLACEHOLDER: Return encrypted unchanged
    # Future: Use Fernet(key).decrypt(encrypted.encode()).decode()
    return encrypted


def generate_encryption_key(password: str, salt: Optional[bytes] = None) -> str:
    """
    Generate encryption key from user password.
    
    Args:
        password: User's password
        salt: Optional salt for key derivation
    
    Returns:
        Base64-encoded encryption key
    
    TODO: Implement using PBKDF2 or Argon2
    """
    # PLACEHOLDER
    # Future: Use PBKDF2 or Argon2 to derive key from password
    return "placeholder_key"


# Example of future implementation:
"""
from cryptography.fernet import Fernet
from cryptography.hazmat.primitives import hashes
from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2
import base64
import os

def generate_encryption_key(password: str, salt: Optional[bytes] = None) -> str:
    if salt is None:
        salt = os.urandom(16)
    
    kdf = PBKDF2(
        algorithm=hashes.SHA256(),
        length=32,
        salt=salt,
        iterations=100000,
    )
    key = base64.urlsafe_b64encode(kdf.derive(password.encode()))
    return key.decode()

def encrypt_content(plaintext: str, encryption_key: str) -> str:
    f = Fernet(encryption_key.encode())
    encrypted = f.encrypt(plaintext.encode())
    return encrypted.decode()

def decrypt_content(encrypted: str, encryption_key: str) -> str:
    f = Fernet(encryption_key.encode())
    decrypted = f.decrypt(encrypted.encode())
    return decrypted.decode()
"""
