pyproject.toml
[tool.poetry]
name = "vaulto-note"
version = "0.1.0"
description = "Secure notes application backend"
authors = ["Antigravity <antigravity@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.10.0"
asyncpg = "^0.31.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
httpx = "^0.24.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.1"
python-multipart = "^0.0.6"
eth-account = "^0.8.0"
python-dotenv = "^1.0.0"
setuptools = "^75.0.0"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
